+++ src/types/index.ts
@@ -1,3 +1,2 @@
 export * from './pipe'
-export * from './map'
 export * from './util'
